Make artifact discovery available to Vercel SSR
