+++ server/package-file-artifacts.d.ts
@@ -1,0 +1,14 @@
+export type PackageFileArtifactPaths = {
+  pcbSvgPath?: string
+  schematicSvgPath?: string
+  circuitJsonPath?: string
+}
+
+type PackageFileLike = {
+  file_path?: string | null
+}
+
+export declare const getPackageFileArtifactPaths: (
+  selectedFilePath: string,
+  packageFiles?: PackageFileLike[],
+) => PackageFileArtifactPaths
